Updated TCR
Added 1 variable to VUEX Strage that starts with TCR.

Also noticed a few problems with Wonder.io which I'll communicate to Aaron.

The pages are no longer being displayed correctly - they are overlapping with the points on the bottom.

Also the scores are no longer saving as well. I'm wondering if it's a security feature in VUE?

I don't believe this TCR project has anything to do with either.

Also note,, I'm uploading videos which I'm sure will be taking up space.